services/investment-brain/pyproject.toml
[project]
name = "investment-brain"
version = "0.1.0"
description = "XXT-AGENT Investment AI Agent System — LangGraph Multi-Agent Orchestrator"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "langgraph>=0.3.0",
    "langchain-core>=0.3.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "orjson>=3.10.0",
    "redis>=5.2.0",
    "numpy>=2.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
